Organize button components
refs #77
