+++ app/page.tsx
@@ -1,4 +1,4 @@
-import Button from '@/components/button'
+import ButtonFilled from '@/components/buttons/filled'
 import Search from '@/components/search'
 import { Metadata } from 'next'
 import dynamic from 'next/dynamic'
@@ -37,7 +37,7 @@ export default function Home() {
       </div>
       {/* Button */}
       <div className='grid justify-center'>
-        <Button text='View more listings' dataCyBtn='btn-view-listings' />
+        <ButtonFilled text='View more listings' dataCyBtn='btn-view-listings' />
       </div>
     </div >
   )
