--- components/button/index.tsx
@@ -1,23 +1,0 @@
-import { MouseEventHandler } from "react";
-
-interface ButtonProps {
-    /**
-     * Button text
-     */
-    text: string;
-    /**
-     * Button action when clicked
-     */
-    onClick?: MouseEventHandler<HTMLButtonElement>;
-    dataCyBtn?: string;
-}
-
-export default function Button(props: ButtonProps) {
-    return (
-        <button className="bg-amber-400 rounded-full px-10 py-4 transition-all hover:shadow-md"
-            data-cy={props.dataCyBtn ?? "button"}
-            onClick={props.onClick}>
-            {props.text ?? "Button"}
-        </button>
-    );
-}
